Moved all tooltip code to its own file
