+++ lib/ts/stacks.ts
@@ -80,68 +80,4 @@ namespace Stacks {
     export function addController(name: string, controller: ControllerDefinition) {
         application.register(name, createController(controller));
     }
-
-    export interface TooltipOptions {
-        placement: string;
-        classNames: string;
-    }
-
-    export function setTooltipHtml(element: Element, html: string, options?: TooltipOptions) {
-        setTooltip(element, popover => popover.innerHTML = html, options);
-    }
-
-    export function setTooltipText(element: Element, text: string, options?: TooltipOptions) {
-        setTooltip(element, popover => popover.textContent = text, options);
-    }
-
-    var generatedTooltipIdNumber = 0;
-
-    function setTooltip(element: Element, fn: (popover: HTMLElement) => void, options?: TooltipOptions) {
-
-        var controllerText = element.getAttribute("data-controller") || "";
-        if (!/(^|\s)s-tooltip(\s|$)/.test(controllerText)) {
-            element.setAttribute("data-controller", controllerText + " s-tooltip");
-        }
-
-        var popoverId = element.getAttribute("aria-describedby");
-        if (!popoverId) {
-            popoverId = "--stacks-s-tooltip-" + (generatedTooltipIdNumber++);
-            element.setAttribute("aria-describedby", popoverId);
-        }
-
-        if (options && options.placement) {
-            element.setAttribute("data-s-tooltip-placement", options.placement);
-        }
-
-        if (options && options.classNames) {
-            element.setAttribute("data-s-tooltip-class-names", options.classNames);
-        }
-
-        var popover = document.getElementById(popoverId);
-        if (!popover) {
-            popover = document.createElement("div");
-            popover.id = popoverId;
-            popover.className = "s-popover s-popover__tooltip";
-            popover.setAttribute("aria-hidden", "true");
-            popover.setAttribute("role", "tooltip");
-
-            var parentNode = element.parentNode;
-            if (parentNode) {
-                // insertBefore inserts at end if element.nextSibling is null.
-                parentNode.insertBefore(popover, element.nextSibling);
-            } else {
-                document.body.appendChild(popover);
-            }
-        }
-
-        const arrow = popover.querySelector(".s-popover--arrow");
-
-        fn(popover);
-
-        if (arrow) {
-            popover.appendChild(arrow);
-        } else {
-            popover.insertAdjacentHTML("beforeend", `<div class="s-popover--arrow"></div>`);
-        }
-    }
 }
